chore: remove stats dependency on database; add konsist check to prevent feature layer depending on data layer
closes #460

+++ data/repository/src/commonMain/kotlin/ly/david/musicsearch/data/repository/area/AreasListRepositoryImpl.kt
@@ -37,7 +37,7 @@ class AreasListRepositoryImpl(
         )
     }
 
-    override fun observeCountOfAllAreas(): Flow<Long> {
+    override fun observeCountOfAllAreas(): Flow<Int> {
         return areaDao.observeCountOfAllAreas()
     }
