+++ data/repository/src/commonMain/kotlin/ly/david/musicsearch/data/repository/series/SeriesListRepositoryImpl.kt
@@ -37,7 +37,7 @@ class SeriesListRepositoryImpl(
         )
     }
 
-    override fun observeCountOfAllSeries(): Flow<Long> {
+    override fun observeCountOfAllSeries(): Flow<Int> {
         return seriesDao.observeCountOfAllSeries()
     }
