+++ shared/domain/src/commonMain/kotlin/ly/david/musicsearch/shared/domain/release/ReleasesListRepository.kt
@@ -12,5 +12,5 @@ interface ReleasesListRepository {
         listFilters: ListFilters,
     ): Flow<PagingData<ReleaseListItemModel>>
 
-    fun observeCountOfReleases(browseMethod: BrowseMethod?): Flow<Long>
+    fun observeCountOfReleases(browseMethod: BrowseMethod?): Flow<Int>
 }
